Abilites Pt. 1
Ich will weinen

Added new Player Input System
Added Abilities
Added Player Input Component to Player
Added testing for Inventory Pieces
Added Add and Subtract Stats for Player Inventory
Added Option to Unequip item
Tried adding Dash, currently bugged tho :(
Added Input Package

+++ BangerProjekt/Assets/Scripts/InventoryScripts/Item.cs
@@ -8,15 +8,15 @@ using UnityEngine.EventSystems;
 
 
 [CreateAssetMenu]
-public class Item : ScriptableObject // Creating a item class for the standart item
+public class Item : ScriptableObject // Creating a item class for the standard item
 {   //weapon 
     public float fireRate;
 
-    public float damage;
+    public int damage;
     //armour
     public float defense;
 
-    public float healthBonus;
+    public int healthBonus;
     //item
     public int ID;
     public Enums.SlotTag itemTag;
